+++ Assets/Scenes/PlayerControllerTest.unity
@@ -149,7 +149,7 @@ MeshCollider:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 19195598}
-  m_Material: {fileID: 13400000, guid: 8b7bdaa1d88022841b08bd90d0032bd5, type: 2}
+  m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 4
@@ -250,10 +250,10 @@ CapsuleCollider:
   m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_Enabled: 1
-  m_Radius: 0.5000001
+  m_Radius: 0.5
   m_Height: 2
   m_Direction: 1
-  m_Center: {x: 0.000000059604645, y: 0, z: -0.00000008940697}
+  m_Center: {x: 0, y: 0, z: 0}
 --- !u!23 &48520419
 MeshRenderer:
   m_ObjectHideFlags: 0
@@ -312,7 +312,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 48520417}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -2, y: 2, z: -2}
+  m_LocalPosition: {x: 2.5, y: 2, z: 2.5}
   m_LocalScale: {x: 1, y: 2.5, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -345,7 +345,7 @@ MeshCollider:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 163994841}
-  m_Material: {fileID: 13400000, guid: 8b7bdaa1d88022841b08bd90d0032bd5, type: 2}
+  m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 4
@@ -443,7 +443,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 354092393}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 1.98, y: 4.41, z: -1.21}
+  m_LocalPosition: {x: 3.5, y: 5, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
@@ -472,18 +472,17 @@ MonoBehaviour:
   jellyCore: {fileID: 1102723987}
   jellyCoreCollider: {fileID: 1102723988}
   directionIndicator: {fileID: 1896820777}
-  movementSpeed: 6
-  groundDrag: 5
-  airDrag: 1
+  movementSpeed: 5
+  groundDrag: 2.5
+  climbDrag: 5
+  airDrag: 0.5
   airbornMovementMultiplier: 0.25
   climbDirectionGravity: 5
   jumpStrength: 8
   jumpCooldown: 0.15
   skinWidthBuffer: 0.1
-  objectBeingClimbed: {fileID: 0}
-  climbDirection: {x: 0, y: 0, z: 0}
-  gravityDirection: {x: 0, y: -1, z: 0}
   currentMovementState: 4
+  objectBeingClimbed: {fileID: 0}
 --- !u!54 &354092404
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -499,7 +498,7 @@ Rigidbody:
   m_IsKinematic: 0
   m_Interpolate: 2
   m_Constraints: 112
-  m_CollisionDetection: 1
+  m_CollisionDetection: 0
 --- !u!1 &616053580
 GameObject:
   m_ObjectHideFlags: 0
@@ -610,7 +609,7 @@ MeshCollider:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 630356270}
-  m_Material: {fileID: 13400000, guid: 8b7bdaa1d88022841b08bd90d0032bd5, type: 2}
+  m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 4
@@ -675,7 +674,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 630356270}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -6.64, y: 4.13, z: -6.78}
+  m_LocalPosition: {x: 7.25, y: 4.25, z: 7.25}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -806,7 +805,7 @@ MeshCollider:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 749379760}
-  m_Material: {fileID: 13400000, guid: 8b7bdaa1d88022841b08bd90d0032bd5, type: 2}
+  m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 4
@@ -2307,7 +2306,7 @@ MeshCollider:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1914388886}
-  m_Material: {fileID: 13400000, guid: 8b7bdaa1d88022841b08bd90d0032bd5, type: 2}
+  m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 4
